Hosting files for project planning.
I'm hosting them here so that the code will work in any environment that can access the github url. This repository is public so it should not be a problem. The only reason I am hosting here is because I don't want to create a new repository just for this.